Shadows intensity depends on opacity & fix ft_atof

+++ libft/srcs/ftb_clamp.c
@@ -6,11 +6,11 @@
 /*   By: tmilon <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/30 17:54:13 by tmilon            #+#    #+#             */
-/*   Updated: 2017/11/30 17:55:29 by tmilon           ###   ########.fr       */
+/*   Updated: 2018/06/08 11:17:45 by tmilon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ftb_clamp(int toclamp, int min, int max)
+double	ftb_clamp(double toclamp, double min, double max)
 {
 	if (toclamp > max || toclamp < min)
 		return (toclamp > max ? max : min);
